Making names consistent
